admin 페이지 작성.

+++ WebContent/admin/admin_book_comment.jsp
@@ -11,37 +11,30 @@
 <!-- head start -->
 <head>
 <%@include file="/include/head.jsp" %>
+<style>
+.page_href{
+	display: unset;
+}
+</style>
 <script>
-$(function(){
-	getList();
-});
-
-//리스트 받아 옴
-function getList(){
+function deleteMember(member_id){
+	if(!confirm("삭제 하시겠습니까?")){
+		return
+	}
+	
 	$.ajax({
-		url:"/admin/book_comment",
-		type:"get",
-		success:function(result){
-			viewList(result);
-		}
-	});
-}
-
-// 리스트 출력
-function viewList(json){
-	var con = $("container");
-}
-
-// 삭제
-function bookCommentDelete(id){
-	$.ajax({
-		url:"/admin/book_comment/"+id,
+		url:"/admin/member/"+member_id,
 		type:"delete",
-		success:function(){
-			getList();
+		success:function(result){
+			var json = JSON.parse(result);
+			if(json.resultCode == 0){
+				alert(json.msg);
+			}
+			location.reload(true);
 		}
-	});
+	})
 }
+
 
 </script>
 <title>관리자용 페이지 - 책 코멘트 관리</title>
@@ -79,7 +72,7 @@ function bookCommentDelete(id){
 				   <% int num = pager.getNum();%>
            		   <% int curPos = pager.getCurPos(); %>
            		   <%for(int i=0; i<pager.getPageSize(); i++){ %>
-           		   <%if(num<1); break; %>
+           		   <%if(num<1) break; %>
            		    <%BookComment comment = commentList.get(i); %>
            		    <tr>
            		    	<td><%=num-- %></td>
@@ -95,12 +88,20 @@ function bookCommentDelete(id){
            		   
            		   <tr>
 						<td colspan="6">
-							[prev] 
+							<%if(pager.getFirstPage()-1>0){ %>
+								<a class="page_href" href="/admin/book_comment/page?currentPage=<%=pager.getFirstPage()-1%>">[prev]</a>
+							<%}else{ %>
+								<a class="page_href" href="javascript:alert('첫페이지 입니다.')">[prev]</a>
+							<%} %> 
 							<%for(int i=pager.getFirstPage(); i<=pager.getLastPage(); i++){ %>
 								<%if(i>pager.getTotalPage()) break; %>
-								[<%=i %>]
+								<a class="page_href" href="/admin/book_comment/page?currentPage=<%=i%>">[<%=i %>]</a>
 							<%} %>
-							[next]
+							<%if(pager.getLastPage()+1<pager.getTotalPage()) {%>
+								<a class="page_href" href="/admin/book_comment/page?currentPage=<%=pager.getLastPage()+1%>">[next]</a>
+							<%}else{ %>
+								<a class="page_href" href="javascript:alert('마지막 페이지입니다.')">[next]</a>
+							<%} %>
 							
 						</td>
 					</tr>
